move some more out of trace_ctx
